# src/swagger-api/documentacao-swagger-api.json
{
  "openapi":"3.0.1",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8083",
      "description":"Generated server url"
    }
  ],
  "paths":{
    "/api/beneficiarios/{beneficiarioId}":{
      "put":{
        "tags":[
          "beneficiario-controller"
        ],
        "operationId":"atualizarBeneficiario",
        "parameters":[
          {
            "name":"beneficiarioId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BeneficiarioDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"object"
                }
              }
            }
          }
        }
      },
      "delete":{
        "tags":[
          "beneficiario-controller"
        ],
        "operationId":"removerBeneficiario",
        "parameters":[
          {
            "name":"beneficiarioId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"object"
                }
              }
            }
          }
        }
      }
    },
    "/api/beneficiarios":{
      "post":{
        "tags":[
          "beneficiario-controller"
        ],
        "operationId":"cadastrarBeneficiario",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BeneficiarioDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/Beneficiario"
                }
              }
            }
          }
        }
      }
    },
    "/api/beneficiarios/{beneficiarioId}/documentos":{
      "get":{
        "tags":[
          "beneficiario-controller"
        ],
        "operationId":"listarDocumentosPorBeneficiario",
        "parameters":[
          {
            "name":"beneficiarioId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"array",
                  "items":{
                    "$ref":"#/components/schemas/Documento"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "BeneficiarioDTO":{
        "type":"object",
        "properties":{
          "nome":{
            "type":"string"
          },
          "telefone":{
            "type":"string"
          },
          "dataNascimento":{
            "type":"string",
            "format":"date"
          },
          "dataInclusao":{
            "type":"string",
            "format":"date"
          },
          "dataAtualizacao":{
            "type":"string",
            "format":"date"
          },
          "documentos":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/DocumentoDTO"
            }
          }
        }
      },
      "DocumentoDTO":{
        "type":"object",
        "properties":{
          "tipoDocumento":{
            "type":"string"
          },
          "descricao":{
            "type":"string"
          },
          "dataInclusao":{
            "type":"string",
            "format":"date"
          },
          "dataAtualizacao":{
            "type":"string",
            "format":"date"
          }
        }
      },
      "Beneficiario":{
        "type":"object",
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          },
          "nome":{
            "type":"string"
          },
          "telefone":{
            "type":"string"
          },
          "dataNascimento":{
            "type":"string",
            "format":"date"
          },
          "documentos":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/Documento"
            }
          }
        }
      },
      "Documento":{
        "type":"object",
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          },
          "tipoDocumento":{
            "type":"string"
          },
          "descricao":{
            "type":"string"
          },
          "beneficiario":{
            "$ref":"#/components/schemas/Beneficiario"
          }
        }
      }
    }
  }
}